% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{lfp}
\alias{lfp}
\title{Female labor force participation data.}
\format{
A data frame with 753 rows and 9 variables:
\describe{
  \item{lfp}{binary indicator for participating in the labor force (=1) or not (=0)}
  \item{intercept}{intercept}
  \item{k5}{number of children 5 years old or younger}
  \item{k618}{number of children 6 to 18 years old}
  \item{age}{in years}
  \item{wc}{wife went to college dummy}
  \item{hc}{husband went to college dummy}
  \item{lwg}{log expected wage rate; for women in the labor force, the actual wage rate; for women not in the labor force, an imputed value based on the regression of \code{lwg} on the other variables}
  \item{inc}{family income exclusive of wife's income}
}
}
\source{
Data taken from 'carData' package. Also known as the 'Mroz' dataset. Mroz, T. A. (1987) The sensitivity of an empirical model of married women's hours of work to economic and statistical assumptions. Econometrica 55, 765-799.
}
\usage{
lfp
}
\description{
A dataset containing socio-economic characteristics as well as a labor
force participation dummy for 753 married women from the panel study
of income dynamics.
}
\keyword{datasets}
