% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{program}
\alias{program}
\title{Students program choices.}
\format{
A data frame with 200 rows and 5 variables:
\describe{
  \item{program}{a vector of program choices}
  \item{intercept}{an intercept term}
  \item{female}{binary indicator for female students}
  \item{ses}{socioeconomic status, 1 is lowest}
  \item{write}{writing score of student}
}
}
\source{
Original dataset is known as 'hsbdemo' and has been sourced from \url{https://stats.oarc.ucla.edu/stat/data/hsbdemo.dta}.
}
\usage{
program
}
\description{
A dataset containing the choice among general program, vocational program and academic program for 200 high school students as well as some explanatory
variables.
}
\keyword{datasets}
