% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMMARY.R
\name{summary.UPG.Logit}
\alias{summary.UPG.Logit}
\title{Estimation results and tables for UPG.Logit objects}
\usage{
\method{summary}{UPG.Logit}(
  object = NULL,
  ...,
  q = c(0.025, 0.975),
  names = NULL,
  digits = 2,
  include = NULL,
  table = NULL,
  cap = NULL
)
}
\arguments{
\item{object}{an object of class \code{UPG.Logit}.}

\item{...}{other summary parameters.}

\item{q}{a numerical vector of length two holding the posterior quantiles to be extracted. Default are 0.025 and 0.975 quantiles.}

\item{names}{a character vector indicating names for the variables used in the output.}

\item{digits}{number of digits to be included in output. Last digit will be rounded using \code{round}.}

\item{include}{can be used to summarize and tabulate only a subset of variables. Specificy the columns of X that should be kept in the plot. See examples for further information.}

\item{table}{can be used to return a LaTeX table (\code{'latex'}), a Word table (\code{'pandoc'}) and HTML tables (\code{'html'}) via \code{knitr}. Include package "booktabs" in LaTeX preamble for LaTeX tables.}

\item{cap}{character vector that can be used to specify the table caption is returned.}
}
\value{
Returns a \code{knitr_kable} object containing the summary table.
}
\description{
\code{summary} generates a summary of estimation results for \code{UPG.Logit} objects. Point estimates, estimated standard deviation as well as credible intervals for each variable are tabulated. In addition, an indicator quickly shows whether the credible interval includes zero or not. In addition, LaTeX, HTML and pandoc tables can be quickly generated via \code{knitr}.
}
\examples{
\donttest{
# estimate a logit model using example data
library(UPG)
data(lfp)
y = lfp[,1]
X = lfp[,-1]
results.logit = UPG(y = y, X = X, type = "logit", verbose=TRUE)

# basic summary of regression results
summary(results.logit)

# generate a LaTeX table with subset of variables and custom names
summary(results.logit,
        include=c(1,3),
        names=c("V. kept 1", "V. kept 3"),
        table="latex")
}
}
\seealso{
\code{\link{plot.UPG.Logit}} to plot the results of a discrete choice model from an \code{UPG.Logit} object.
\code{\link{predict.UPG.Logit}} to predict probabilities from a discrete choice model from an \code{UPG.Logit} object.
\code{\link{coef.UPG.Logit}} to extract coefficients from an \code{UPG.Logit} object.
}
\author{
Gregor Zens
}
