% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Grouped Titanic survival data.}
\format{
A data frame with 78 rows and 6 variables:
\describe{
  \item{survived}{number of passengers that survived}
  \item{total}{number of total passengers}
  \item{intercept}{an intercept}
  \item{pclass}{passenger class (1 highest, 3 lowest)}
  \item{female}{dummy for females}
  \item{age.group}{age group indicator (0-5yrs, 5-10yrs, ...)}
}
}
\source{
Data originally sourced from \url{https://web.stanford.edu/class/archive/cs/cs109/cs109.1166/stuff/titanic.csv}. See also
\url{https://towardsdatascience.com/the-binomial-regression-model-everything-you-need-to-know-5216f1a483d3}.
}
\usage{
titanic
}
\description{
A dataset containing the number of survivals and the total number of persons by
passenger class, age group and gender.
}
\keyword{datasets}
