% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGNOSTICS.R
\name{UPG.Diag.Binomial}
\alias{UPG.Diag.Binomial}
\title{MCMC Diagnostics for \code{UPG.Binomial} objects}
\usage{
UPG.Diag.Binomial(object = NULL)
}
\arguments{
\item{object}{an object of class \code{UPG.Binomial}.}
}
\value{
Returns a list containing effective sample size, effective sampling rate and inefficiency factors for each coefficient.
}
\description{
\code{UPG.Diag.Binomial} computes inefficiency factors, effective sample size and effective sampling rate based on the posterior distributions in a \code{UPG.Binomial} object.
}
\author{
Gregor Zens
}
