% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGNOSTICS.R
\name{UPG.Diag.Logit}
\alias{UPG.Diag.Logit}
\title{MCMC Diagnostics for \code{UPG.Logit} objects}
\usage{
UPG.Diag.Logit(object = NULL)
}
\arguments{
\item{object}{an object of class \code{UPG.Logit}.}
}
\value{
Returns a list containing effective sample size, effective sampling rate and inefficiency factors for each coefficient.
}
\description{
\code{UPG.Diag.Logit} computes inefficiency factors, effective sample size and effective sampling rate based on the posterior distributions in an \code{UPG.Logit} object.
}
\author{
Gregor Zens
}
