% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGNOSTICS.R
\name{UPG.Diag.MNL}
\alias{UPG.Diag.MNL}
\title{MCMC Diagnostics for \code{UPG.MNL} objects}
\usage{
UPG.Diag.MNL(object = NULL)
}
\arguments{
\item{object}{an object of class \code{UPG.MNL}.}
}
\value{
Returns a list containing effective sample size, effective sampling rate and inefficiency factors for each coefficient.
}
\description{
\code{UPG.Diag.MNL} computes inefficiency factors, effective sample size and effective sampling rate based on the posterior distributions in a \code{UPG.MNL} object.
}
\author{
Gregor Zens
}
