% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGNOSTICS.R
\name{UPG.Diag.Probit}
\alias{UPG.Diag.Probit}
\title{MCMC Diagnostics for UPG.Probit objects}
\usage{
UPG.Diag.Probit(object = NULL)
}
\arguments{
\item{object}{an object of class \code{UPG.Probit}.}
}
\value{
Returns a list containing effective sample size, effective sampling rate and inefficiency factors for each coefficient.
}
\description{
\code{UPG.Diag.Probit} computes inefficiency factors, effective sample size and effective sampling rate based on the posterior distributions in an \code{UPG.Probit} object.
}
\author{
Gregor Zens
}
